requirements.txt
requests
python-telegram-bot[job-queue]
aiohttp
aioping
aiofiles
certifi